Refactor: Consolidate protocol_schema.py into protocol/ramses.py
