+++ src/ramses_rf/parsers/decoder.py
@@ -9,7 +9,7 @@ import re
 from abc import ABC, abstractmethod
 from typing import Any
 
-from ramses_rf.protocol_schema import (
+from ramses_rf.protocol.ramses import (
     CODE_IDX_ARE_COMPLEX,
     CODE_IDX_ARE_NONE,
     CODE_IDX_ARE_SIMPLE,
@@ -210,7 +210,7 @@ class _LegacyMessage:
         return self._idx_
 
     def _pkt_idx(self) -> bool | str | None:
-        """Extract the exact index leveraging protocol_schema definitions.
+        """Extract the exact index leveraging protocol.ramses definitions.
 
         :return: String or boolean representation if index parsing passes constraints.
         :raises exc.PacketPayloadInvalid: If structural legacy checks fail validations.
